Rewrite SDP body parser. By definition in the RFC4566, the SDP body has very simple fixed syntax, so that complexity of our internal representation modelled after more complex SIP objects is not justifyed. This brings up to 20-25% speed increase on parse/__str__() operations, but what also important makes using the SDP classes easier and allows us to implement more features easily, which otherwise would be too complex to do penalizing execution time.
Implement one such feature: if there is only one media description
in the message, move connection parameters from it to the session
description, since some of the devices that don't implement RFC4566
properly might break.

The other important feature to be implemented later on is automatic
elimination of the a=rtpmap entries for media types that are no longer
present in the media description.